Refactor: Razor cuts - remove bloat and simplify assembler
