+++ src/tools/mod.rs
@@ -20,11 +20,6 @@ pub mod cache;
 pub mod cli;
 pub mod dictionary;
 pub mod highlight;
-#[cfg(feature = "nova")]
-pub mod mentor;
-#[cfg(feature = "nova")]
-pub mod mosaic;
-pub mod narrator;
 pub mod repl;
 pub mod report;
 pub mod runner;
